Several Enhancements
+ Separated classes into files.
+ Fixed `dictionary` key being popped when `metadata()` method is being
  called.
